Make Request work better
Since properties can be non-'static the lifetime bounds for satisfy and satisfy_with must also be very generic.
Until now this was pinned to the lifetime of the request itself, but that is somewhat wrong in that the value only needs to live for the call to `satisfy` itself.

Additionally satisfy_with was limited to SizedProperty where it makes the most sense.